[IMP] website: review `s_rating` snippet design
This commit aims to review the `s_rating` snippet design to make it more
in line with the new snippets design.

With the new snippets introduced in Odoo 18 and all the new possibilities
unlocked for the user, it made perfectly sense to review the design of
the `s_rating` snippet to make it fit as much scenario as possible.

task-4306631

Co-authored-by: bvr-odoo <bvr@odoo.com>