[FIX] sale{,_stock}: further update SO if prod add to delivery
**Current behavior:**
Adding a product which is invoiced based on delivered
quantities to the delivery of a sale which has already been
confirmed, then validating the delivery, does not reflect the
change in price (which will be invoiced) on the sale order.

**Expected behavior:**
The sale order should show the actual amount of money to be
charged to the orderer.

**Steps to reproduce:**
1. Create a stored product with some price and change its
invoicing policy to 'delivery'

2. Create a new sale order, add 1 of the created product,
then confirm it

3. Open the delivery, add another line to it for the same
created product, fill in the quantities for each move, then
validate the picking

4. Check the sale order line to see that, while the price unit
is populated with the product's price, the subtotal/total fields
are not

**Cause of the issue:**
Order lines added in this manner have `product_uom_qty == 0`
which will zero-out the price computation.

**Fix:**
Use the maximum between `qty_delivered` and `product_uom_qty`
when a SOL's product is invoiced on delivered quantities.

Thus, the order will always reflect the actual price to be
charged in the invoice (on the invoice, it will still be
indicated which quantities come from the order and which were
added directly on the delivery).

opw-4210056