[FIX] portal: fix customer information form if no country set
Having no country on a portal user would cause attempting to edit
personal information through the portal to have an internal server
error.

Added a check to country_id to ensure that the country_id is a numerical
value before casting to an int to avoid this internal server error.

Allowed the customer to set a country if there is no country set. Also
added value checking on the post request to ensure that user would not
experience incorrect greying out of the country_id when submitting with
errors.

opw-4309159