[FIX] web_editor,website: fixing carousel traceback and transition issue
Issue 1:

Step to reproduce:
1. Drag and Drop a Carousel Snippet.
2. Quickly add slides, remove slides, add slides, remove slides... etc
3. Traceback Occurs.

Issue:

We can't find any editor whose DOM target is still inside the document.
So, when we access "editors[0]" then it throws a traceback.

Solution:

To resolve this issue, we added a check for "editors.length" before
accessing "editors[0]". This prevents the traceback from occurring.

Issue 2:

Step to reproduce:
1. Drag and Drop a Carousel Snippet.
2. Change the Transition from Slide to Fade.
3. Remove all the background images and color.
4. The transition malfunctions

Issue:

The carousel transition CSS currently has an ease-in and ease-out
duration of "0.6s", causing the old text to be removed and the new text
to be inserted simultaneously. This setup results in a visual
malfunction in the transition.

Solution:

To solve this issue we are increasing ease-in time to "3s" and redusing
ease-out to "0.2s".