[IMP] mrp: Simulate planning in MO Overview
Consider workcenters availability in MO Overview like in BoM Overview.
Use 'button_plan' to exactly plan workorders but do not register any
related information.
For this we use a savepoint to rollback the transaction
and a fake sequence to not have any hole in the mo's numbering.

See odoo/enterprise#73385