[FIX] hr: update `distance_from_work` on `km_home_work` changes
To reproduce, sign a new contract in the salary configurator with
a different home to work distance. Check the employee's private
info and the distance is not updated.

The issue happens because the salary configurator works on the
field `km_home_work` while there is also `distance_home_work`
that works with either miles or kilometers.
An inverse was introduced in:  https://github.com/odoo/odoo/pull/178160
but it used the wrong field when computing `distance_from_work`.

task-4306551