[IMP] mrp: ensure correct computation of byproducts
Currently, when we change the quantity for a
MO, all byproduct quantities are updated regardless
of whether the Produced checkbox is picked or not.
With this commit, the byproduct quantities will be
updated just when the Produced checkbox is unpicked.
Additionally, the checkbox will no longer be automatically
picked when the quantity value is modified.