[FIX] web: notify user when a required field is left empty
Before this commit:
In the list view, if a user clears a required many2many field and attempts to
save the record, the record is saved without notifying the user about the
invalid field.

After this commit:
Now, if a user clears a required field and tries to save the record, then the
record does not get saved, and the user receives a notification indicating the
field is invalid.

Task-3981155