[FIX] web_editor, website: insert link in mass_mailing
Issue:
======
We can't insert a link with `/url` in mass mailing

Steps to reproduce the issue:
=============================
- Install website and email marketing
- Create a new mass mailing
- Add a button using `/button`
- Add a label
- Write `/` in url and choose any option make sure to click on it using
  the mouse
- Save
- The link isn't inserted

Origin of the issue:
====================
When we click to select an item, the event `mousedown` is triggered on
the document which will force a wysiwyg blur. So the selection is no
more in the editable and we can't insert the link.

Solution:
=========
To avoid the blur, we add the selector of the autocomplete to the
selectors that delay the blur.

opw-4283325