[IMP] hw_drivers: use requirements.txt to install python deps
We currently hardcode Python dependencies in a script used to create the image
of the IoT Box.

We now have a `requirements.txt` file that lists all python modules, and that is
included in the checkout: every time an IoT Box checks out the Odoo repository,
it will install the requirements.
We also added a `packages.txt` to store all debian packages.

We will now be able to add modules without releasing a new image.

Task: 4188175