[FIX] pos_repair: Fix quotation from repair
When opening a quotation created from a repair order in the PoS the
quantity of the products would always be 0.

Steps to reproduce:
-------------------
* Create a repair order for whatever product
* Add some product to the list with the "Add" option
* Start and End the reparation
* Create a quotation for the repair order
* Open the quotation in the PoS
> Observation: The quantity of the product in the pos is 0

Why the fix:
------------
We added a new computed field to mark the lines that are linked
to a repair order. Those lines are always marked as delivered and
because of this the quantity in the PoS would always be 0.
So we ignore the delivered quantity for those lines and just take
the quantity left to invoice.

opw-4261097