[IMP] project: alter `doAction` to add additional context
Before this commit, some columns in the list view are hidden according
to the context defined in session key to know which features are enabled
or disabled inside the project shared.

This commit alters the `doAction` to be able to add context defined in
the session to make sure some columns inside the list view of project
sharing are visible only if the project has the right feature enabled.

task-4267190