[FIX] web_editor: fix custom colors in colorpicker
Before this commit, the colorpalette for the background color would
not display custom colors already on the page. The issue appeared in
17.0 and worked in 16.0 before the conversion of wysiwyg to owl.
github.com/#118966

This commit make sure the necessary function is added.

Steps to reproduce the bug:
- Add a snippet
- Set the background / font color of the snippet to a custom color
- Add another snippet
- Open the colorpicker for the text color
(The colorpicker have access to the custom color set above)
- Open the colorpicker for the background color
(The colorpicker does not have access to the custom color set above)

task-3806989