[FIX] hr_timesheet: resolve access right error on creating new task on portal
Steps to reproduce:
- Install project and timesheet app.
- Create a project.
- Create the task with 2 subtasks.
- Share the project with the portal user with edit rights.
- Login from the portal user.
- Open project.
- Open task form view.
- Click on the new button on the left side.
- Add the title and click the save button.
- Access rights error occured related to the timesheet_ids.

Issue:
An error occurred of 'timesheet_ids'. So, a new task will not be create from the
portal even when the user has project edit rights.

Cause:
The field 'timesheet_ids' is not read-only. When creating the task,
'timesheet_ids' is available in create's vals_list, and the portal user does not
have write access to the timesheet. So, an access right error occurred.

Solution:
As the portal user does not have the right to create, edit and delete the task's
timesheet, we can make the field 'timesheet_ids' readonly.

task-3969988