[FIX] web_editor: prevent tab duplication
Steps To Reproduce:
-> Go to Edit mode.
-> Drag & Drop Tabs Snippet.
-> Click in middle of tab's text and press "Enter".
-> Tab is split and and new Tab is created.

Issue Reason:
As tab's element fails isUnbreakable() check for 'Keydown'(Enter) event
which makes it possible to split the element. And issue is not specific
to Tab snippet but in other snippets where anchor tag and elements with
'btn' class or role attribute is 'button'.

Solution:
By modifying checks in isUnbreakable() for handling elements with anchor
tags and elements with 'btn' class or role attribute is 'button'.

This commit solves the issue of tab duplication in Tab Snippet and also
in other snippets where anchor tag and elements with 'btn' class or role
attribute is 'button'.

task-4316648