[FIX] purchase_view: Adjust button layout
In Odoo, an issue araised where the "Expected Arrival" date field was
overlapped when zooming in. This was due to the button
"action_purchase_vendor_delay_report".

To adress this, the related button was encapsulated within a <div>
element. This change ensure that the layout remains consistent and the
elements does not overlap regardless of the zoom level.

opw-4316093