[FIX] sale_project: generated task should be in the project with lowest sequence
1. Let's assume an SO containing the following SOL's products:
- A product that generates a project (or a project & a task)
- A product that generates a task  in no particular project

In that case, at SO confirmation, the second product should generate a task in the project generated by the first product (unless a project was selected on the SO).
After this commit, if the SO generates multiple projects and no project was selected on the SO, we generate the task in the generated project with the lowest sequence.

2. Secondly, in this commit, we fix the following issue:
- Create two SOs having both a service product which creates a project
- Confirm both SOs at the same time
- Both SOs have the same project, while two projects have been generated (one per SO) => Each SO should have the project it has generated.
In other terms, we should make it work when confirming multiple SOs in batch.

task-4244890
version-18.0