[FIX] web: Hoot - side effects between tests
This commit fixes the following issues:

- before this commit, some "timers" could be registered after the final
call to `cancelAllTimers` if they were initiated by callbacks spawned
via Owl's RAF function (which is not affected by Hoot). This commit
waits for an additional animation frame before running `cancelAllTimers`
(while disabling new timers) to ensure they are cleaned up as well;

- previously, errors would try to "interrupt" the current test by
rejecting the promise related to its completion. The issue is that said
promise doesn't affect the test block execution, so any error that tried
to shut down a test would just go to the next one without stopping the
previous. This commit removes that "stopping" effect, so that a test
will either complete its execution or be aborted by an actual block-
scoped error.