[IMP] event, event_sale, pos_event: share state between event and pos
Before the `state` fields was only managed with event_sale. But since we
are able to sell tickets from the POS, we need to share the state
between these modules to have a coherent state.

This commit adds the `state` field to the `event.registration` model
directly in `event` module. This way, the state is shared between
`event`, `event_sale` and `pos_event`.

This commit also add a `oe_stat_button` to the `event.registration`
form view to be able to access to its pos_order directly.

taskId: 4330905