[IMP] conditional_format: improve CF data bar behavior for non-matching range sizes
Before this commit, defining a CF data bar rule based on the value of
another range requires both ranges to have matching sizes, preventing
the CF from being created otherwise. This restriction leads to a poor
user experience when copying and pasting CFs, as the CF does not get
pasted.

This commit removes the restriction, allowing CF creation in a "best effort"
manner when the range sizes do not match. The CF style is therefore applied by
comparing the matching cells.

task-4280720