[IMP] point_of_sale: Do not display lot number that are not available
Only display lot numbers if their forecasted quantity is available. Now when using a 2 step routes (inventory) we only show in PoS the serial numbers that are not already reserved for a stock move.
To do this we have to use the quantity inside `stock.quant` because the previously used quantity (from `stock.lot`) is only updated when the pickup is confirmed.